Fix previous newdle data leaking into newdle creation
When a user visits a specific newdle (`/newdle/{code}/summary`),
some data is automatically prefilled in the `creation`
redux state. If the user goes back to `/mine` and creates a
new newdle, this data is not properly removed.

This commit adds a call to `abortCreation` when newdle summary
is unmounted, which clears the state.