Correctly handle available timeslots when deselecting
Fixes an issue when deselecting an answer and saving
in the `limited slots` mode would make it disabled
even though by deselecting it, it just became available.

The cause was not refetching the updated newdle &
the `available_timeslots`. Refetching is not really feasible
though since there are some `useEffect`s watching for `newdle`
which causes extra fetches.

Workaround is to keep track of the available timeslots locally.