Prevent popup from flickering when hovering a time slot.
This happens when the popup covers a part of the trigger
when it renders which prevents the original hover event
which hides the popup again and so on.

The problem here was the 'x' delete icon of the time slot
which was being covered by the popup. The solution is to make
only the slot itself as the trigger.

Another option would be to use the `hoverable` or `offset`
props of SUI's <Popup />.